# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Autonomous Shopping AI Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductSpecs = { [key: string]: unknown };

export interface Product {
  id: number;
  name: string;
  brand: string;
  category: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  description: string;
  specs?: ProductSpecs;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  isFeatured: boolean;
  tags: string[];
  /** @nullable */
  aiScore?: number | null;
  createdAt?: string;
}

export type ProductInputSpecs = { [key: string]: unknown };

export interface ProductInput {
  name: string;
  brand: string;
  category: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  description: string;
  specs?: ProductInputSpecs;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  isFeatured: boolean;
  tags: string[];
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  icon: string;
  productCount: number;
}

export type RecommendRequestPreferences = { [key: string]: unknown };

export type RecommendRequestConversationHistoryItem = {
  role: string;
  content: string;
};

export interface RecommendRequest {
  query: string;
  /** @nullable */
  budget?: number | null;
  /** @nullable */
  category?: string | null;
  preferences?: RecommendRequestPreferences;
  conversationHistory?: RecommendRequestConversationHistoryItem[];
}

export interface ProductRecommendation {
  product: Product;
  score: number;
  rank: number;
  why: string;
  pros: string[];
  cons: string[];
  bestFor: string;
  notSuitableFor: string;
  reviewSummary: string;
}

export interface RecommendResponse {
  reasoning: string;
  products: ProductRecommendation[];
  /** @nullable */
  budgetAdvice?: string | null;
  accessories: string[];
  followUpQuestions: string[];
}

export interface CompareRequest {
  productIds: number[];
  /** @nullable */
  userNeeds?: string | null;
}

export type CompareResponseComparisonTableItemScores = { [key: string]: unknown };

export type CompareResponseComparisonTableItem = {
  aspect: string;
  scores: CompareResponseComparisonTableItemScores;
};

export interface CompareResponse {
  winner: Product;
  analysis: string;
  comparisonTable: CompareResponseComparisonTableItem[];
  recommendation: string;
}

export interface AiScore {
  productId: number;
  overall: number;
  performance: number;
  price: number;
  reliability: number;
  battery: number;
  popularity: number;
  futureProof: number;
  summary: string;
}

export interface AiMemory {
  id: number;
  sessionId: string;
  /** @nullable */
  budget?: number | null;
  favoriteCategories: string[];
  favoriteBrands: string[];
  /** @nullable */
  preferredPriceRange?: string | null;
  /** @nullable */
  userProfile?: string | null;
  shoppingGoals: string[];
  updatedAt?: string;
}

export interface AiMemoryUpdate {
  /** @nullable */
  budget?: number | null;
  favoriteCategories?: string[];
  favoriteBrands?: string[];
  /** @nullable */
  preferredPriceRange?: string | null;
  /** @nullable */
  userProfile?: string | null;
  shoppingGoals?: string[];
}

export type PricePredictionPriceHistoryItem = {
  date: string;
  price: number;
};

export interface PricePrediction {
  productId: number;
  currentPrice: number;
  predictedTrend: string;
  confidence: number;
  bestTimeToBuy: string;
  analysis: string;
  priceHistory: PricePredictionPriceHistoryItem[];
}

export interface CartItem {
  id: number;
  sessionId: string;
  productId: number;
  quantity: number;
  product: Product;
  addedAt?: string;
}

export interface CartItemInput {
  productId: number;
  quantity: number;
  sessionId: string;
}

export interface WishlistItem {
  id: number;
  sessionId: string;
  productId: number;
  product: Product;
  addedAt?: string;
}

export interface WishlistItemInput {
  productId: number;
  sessionId: string;
}

export interface RecommendationHistory {
  id: number;
  sessionId: string;
  query: string;
  reasoning: string;
  productIds: number[];
  createdAt: string;
}

export interface DashboardStats {
  totalProducts: number;
  totalCategories: number;
  avgAiScore: number;
  recommendationsToday: number;
  cartItems: number;
  wishlistItems: number;
  savedBudget: number;
  topCategory: string;
}

export interface TrendingCategory {
  category: string;
  count: number;
  avgPrice: number;
  avgScore: number;
}

export interface PriceAlert {
  productId: number;
  productName: string;
  originalPrice: number;
  currentPrice: number;
  discount: number;
  /** @nullable */
  imageUrl?: string | null;
}

export type ListProductsParams = {
category?: string;
search?: string;
minPrice?: number;
maxPrice?: number;
brand?: string;
limit?: number;
offset?: number;
};