# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Autonomous Shopping AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface CartItem {
  id: number;
  sessionId: string;
  productId: number;
  quantity: number;
  product: Product;
  addedAt?: string;
}