# lib/api-zod/src/generated/types/pricePrediction.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Autonomous Shopping AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { PricePredictionPriceHistoryItem } from './pricePredictionPriceHistoryItem';

export interface PricePrediction {
  productId: number;
  currentPrice: number;
  predictedTrend: string;
  confidence: number;
  bestTimeToBuy: string;
  analysis: string;
  priceHistory: PricePredictionPriceHistoryItem[];
}