# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Autonomous Shopping AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductSpecs } from './productSpecs';

export interface Product {
  id: number;
  name: string;
  brand: string;
  category: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  description: string;
  specs?: ProductSpecs;
  rating: number;
  reviewCount: number;
  inStock: boolean;
  isFeatured: boolean;
  tags: string[];
  /** @nullable */
  aiScore?: number | null;
  createdAt?: string;
}