# lib/api-zod/src/generated/types/productRecommendation.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Smart Autonomous Shopping AI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductRecommendation {
  product: Product;
  score: number;
  rank: number;
  why: string;
  pros: string[];
  cons: string[];
  bestFor: string;
  notSuitableFor: string;
  reviewSummary: string;
}